moved useHandlePayment to hooks folder
